+++ packages/components/src/interface/index.ts
@@ -16,7 +16,6 @@ export * from './SelectGroup'
 export * from './ListViewSimplified/ListViewSimplified'
 export * from './DataTable/SortAndFilter'
 export * from './SectionDrawer'
-export * from './Divider'
 export * from './SearchTool'
 export * from './ExpandableNotification/ExpandableNotification'
 export * from './Header/AppHeader'
